+++ NEWS.md
@@ -1,5 +1,6 @@
 # MCView 0.2.43
 
+* Added "Highlight selected genes" option to the 'Markers' tab.
 * Fix: changing `force_cell_type` did not update the heatmap when in zoom mode.
 * Fix: show "Enrichment type" toggle when filtering by clipboard, and allow updating genes.
 * Added profiling mode for debug purposes.